Fix compile and tests
